Added debug mode to all resources.
When debugMode POST parameter is set to true (default false) a
custom appender is temporarily added to the root logger to
capture all the log events raised while serving the request.
The resulting capture is then returned in a JSON field of the response.

Changed the logging library from slf4j-simple to logback.

The factory of the JSON responses now writes the field in a fixed order.

Added unit tests of the feature.

Modified the web interface to ask for debug mode and show resulting
logs.